requirements.txt
Paver==1.0.4
PyYAML==3.10
colorlog==1.6
keepassdb==0.2.1
nose==1.2.1
pycrypto==2.6
python-gnupg==0.3.1
unittest2==0.5.1